# apis/docs/swagger.yaml
basePath: /
definitions:
  commons.ApiErrorResponsePayload:
    properties:
      additional_info:
        additionalProperties: true
        type: object
      error:
        type: string
      status:
        type: string
    type: object
  models.User:
    properties:
      _id:
        type: string
      age:
        type: integer
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is a sample API using Echo and Swagger.
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: get details of all users
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.ApiErrorResponsePayload'
      summary: GetUsers
      tags:
      - User Management
    post:
      consumes:
      - application/json
      description: Create a user with name, email, age, and is_Active status
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.ApiErrorResponsePayload'
      summary: CreateUser
      tags:
      - User Management
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user details by user id
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.ApiErrorResponsePayload'
      summary: DeleteUserById
      tags:
      - User Management
    get:
      consumes:
      - application/json
      description: Gets user details by user id such as name, email, status etc.
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/commons.ApiErrorResponsePayload'
      summary: GetUserById
      tags:
      - User Management
swagger: "2.0"